requirements.txt
boto3==1.26.42
argparse
requests==2.28.2
docker==6.0.1
gitdb==4.0.10
GitPython==3.1.30
simple-colors==0.1.5
requests==2.28.2